backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.12
openai==1.0.0 
sqlalchemy
asyncpg
python-multipart